chore(gitignore): mise à jour des fichiers ignorés
- Ajout/Modification des règles d'ignorance
- Ignore libftTester/ et fichiers objets

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: maherraz <maherraz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/24 19:40:11 by maherraz          #+#    #+#             */
-/*   Updated: 2026/01/25 03:52:27 by maherraz         ###   ########.fr       */
+/*   Updated: 2026/02/10 19:16:21 by maherraz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned int	i;
 	unsigned char	*ptr;
