+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: maherraz <maherraz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/24 16:30:31 by maherraz          #+#    #+#             */
-/*   Updated: 2026/01/25 05:00:49 by maherraz         ###   ########.fr       */
+/*   Updated: 2026/02/11 14:36:06 by maherraz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && (s1[i] != '\0' || s2[i] != '\0'))
